@@ -1,3 +1,3 @@
-module conversion
+module github.com/subrajeet-maharana/golang/04conversion
 
-go 1.19
+go 1.22.0
