@@ -1,3 +1,3 @@
-module mytime
+module github.com/subrajeet-maharana/golang/06mytime
 
-go 1.19
+go 1.22.0
