@@ -1,3 +1,3 @@
-module mypointers
+module github.com/subrajeet-maharana/golang/07mypointers
 
-go 1.19
+go 1.22.0
