@@ -1,3 +1,3 @@
-module myarray
+module github.com/subrajeet-maharana/golang/08myarray
 
-go 1.19
+go 1.22.0
