@@ -1,3 +1,3 @@
-module slices
+module github.com/subrajeet-maharana/golang/09myslices
 
-go 1.19
+go 1.22.0
