@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"slices"
 	"sort"
 )
 
@@ -15,7 +16,7 @@ func main() {
 	// fmt.Println(fruitList)
 	// fruitList = append(fruitList[1:])
 	// fmt.Println(fruitList)
-	fruitList = append(fruitList[:3])
+	fruitList = append(fruitList[:3], fruitList...)
 	fmt.Println(fruitList)
 	highScores := make([]int, 4)
 	highScores[0] = 783
@@ -33,6 +34,7 @@ func main() {
 	//remove value from slice based on index
 	var courses = []string{"Web Dev", "App Dev", "Devops", "ML", "SDE", "Blockchain", "HFT"}
 	var index int = 4
-	courses = append(courses[:index], courses[index+1:]...)
+	// courses = append(courses[:index], courses[index+1:]...)
+	courses = slices.Delete(courses, index, index+1)
 	fmt.Println(courses)
 }
