@@ -4,12 +4,14 @@ import "fmt"
 
 func main() {
 	fmt.Println("Welcome to Structs in Golang!")
-	subrajeet := User{"Subrajeet", "subrajeet@go.dev", true, 22}
+	subrajeet := User{"Subrajeet", "subrajeet@go.dev", true, 25}
 	fmt.Println(subrajeet)
 	fmt.Printf("The details of Subrajeet are %+v\n", subrajeet)
 	fmt.Printf("The name is %v and email is %v\n", subrajeet.Name, subrajeet.Email)
 	subrajeet.GetStatus()
 	subrajeet.NewEmail()
+	fmt.Printf("The name is %v and email is %v\n", subrajeet.Name, subrajeet.Email)
+	subrajeet.NewUpdatedEmail()
 	fmt.Printf("The name is %v and email is %v\n", subrajeet.Name, subrajeet.Email)
 }
 
@@ -25,7 +27,14 @@ type User struct {
 func (u User) GetStatus() {
 	fmt.Println("Is user active: ", u.Status)
 }
+
 func (u User) NewEmail() {
+	u.Email = "test@go.dev"
+	fmt.Println("The new email is: ", u.Email)
+}
+
+// This following method takes pointer and changes the actual struct instance
+func (u *User) NewUpdatedEmail() {
 	u.Email = "test@go.dev"
 	fmt.Println("The new email is: ", u.Email)
 }
